[refactor, docs] class별로 파일 분리, class를 생성하는 class를 함수로 변경
