+++ app.js
@@ -6,10 +6,10 @@ const app = express();
 const port = 8080;
 
 app.use(cors());
-// app.use(express.static('public'));
+app.use(express.static('public'));
 
 app.get('/', (req, res) => {
-    // res.header('Content-Type', 'application/json');
+    res.header('Content-Type', 'application/json');
     res.json(data);
 })
