+++ index.html
@@ -7,10 +7,7 @@
     <title>carousel</title>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" />
     <link rel="stylesheet" href="./style.css" />
-    <script type="text/javascript" src="./template.js"></script>
-    <script type="text/javascript" src="./carousel.js"></script>
-    <script type="text/javascript" src="./fetch.js"></script>
-    <script type="text/javascript" src="./config.js"></script>
+    <script type="module" src="./main.js"></script>
 
 </head>
