--- mockData.js
@@ -1,33 +1,0 @@
-const MENU_DATA = [
-    'Hypertext Markup Language',
-    'Cascading Style Sheets',
-    'JavaScript',
-    'React (web framework)'
-]
-
-const CONTENTS_DATA = [
-    {
-        imgUrl: './images/img1.png',
-        titleBox: 'Hypertext Markup Language',
-        title: 'HTML',
-        content: 'Web browsers receive HTML documents from a web server or from local storage and render the documents into multimedia web pages. HTML describes the structure of a web page semantically and originally included cues for the appearance of the document.'
-    },
-    {
-        imgUrl: './images/img2.png',
-        titleBox: 'ascading Style Sheets',
-        title: 'CSS',
-        content: 'CSS is designed to enable the separation of presentation and content, including layout, colors, and fonts. This separation can improve content accessibility, provide more flexibility and control in the specification of presentation characteristics, enable multiple web pages to share formatting by specifying the relevant CSS in a separate .css file, and reduce complexity and repetition in the structural content.'
-    },
-    {
-        imgUrl: './images/img3.png',
-        titleBox: 'Javascript',
-        title: 'Javascript',
-        content: 'Alongside HTML and CSS, JavaScript is one of the core technologies of the World Wide Web. JavaScript enables interactive web pages and is an essential part of web applications. The vast majority of websites use it for client-side page behavior, and all major web browsers have a dedicated JavaScript engine to execute it.'
-    },
-    {
-        imgUrl: './images/img4.png',
-        titleBox: 'React (web framework)',
-        title: 'React',
-        content: 'React can be used as a base in the development of single-page or mobile applications. However React is only concerned with rendering data to the DOM and so creating React applications usually requires the use of additional libraries for state management and routing. Redux and React Router are respective examples of such libraries.'
-    },
-]
