+++ util.js
@@ -1,0 +1,3 @@
+export const $ = (selector, all) => {
+    return all ? document.querySelectorAll(selector) : document.querySelector(selector);
+}
